Cambio al tipo de dato en DTO

+++ CERVERICA/DTO/Recetas/PasosInsertDto.cs
@@ -4,6 +4,6 @@ namespace CERVERICA.Dtos
     {
         public string Descripcion { get; set; }
         public int Orden { get; set; }
-        public int Tiempo { get; set; }
+        public double Tiempo { get; set; }
     }
 }
